+++ src/TelescopeSourceGenerator/Analyzers/TelescopeIncrementalGenerator.cs
@@ -225,159 +225,3 @@ public class TelescopeIncrementalGenerator : IIncrementalGenerator
         MarkExportAttributeParseResult ReturnFalse() => new MarkExportAttributeParseResult(false, default, default);
     }
 }
-
-class AttributedTypesExportGenerator
-{
-    public string Generate(ImmutableArray<MarkClassParseResult> markClassCollection, CancellationToken token)
-    {
-        var exportedInterfaces = new List<string>();
-        var exportedMethodCodes = new List<string>();
-
-        foreach (var markClassGroup in markClassCollection.GroupBy(t => t.MarkExportAttributeParseResult))
-        {
-            token.ThrowIfCancellationRequested();
-
-            var markExportAttributeParseResult = markClassGroup.Key;
-            /*
-             * ExportedTypeMetadata<TBaseClassOrInterface, TAttribute>[] ICompileTimeTypesExporter<baseClassOrInterfaceName, attributeName>.ExportTypes()
-             * {
-             *    return new ExportedTypeMetadata<TBaseClassOrInterface, TAttribute>[]
-             *    {
-             *       new ExportedTypeMetadata<TBaseClassOrInterface, TAttribute>(typeof(type), () => new {type}())
-             *    }
-             * }
-             */
-            var baseClassOrInterfaceName =
-                TypeInfoToFullName(markExportAttributeParseResult.BaseClassOrInterfaceTypeInfo);
-            var attributeName = TypeInfoToFullName(markExportAttributeParseResult.AttributeTypeInfo);
-
-            var exportedItemList = new List<string>();
-
-            foreach (var markClassParseResult in markClassGroup.Select(t => t.ClassParseResult))
-            {
-                // new ExportedTypeMetadata<TBaseClassOrInterface, TAttribute>(typeof(type), () => new {type}())
-                var typeName = TypeSymbolToFullName(markClassParseResult.TypeInfo);
-
-                var itemCode = @$"new ExportedTypeMetadata<{baseClassOrInterfaceName}, {attributeName}>(typeof({typeName}), () => new {typeName}())";
-                exportedItemList.Add(itemCode);
-            }
-
-            var arrayExpression = $@"new ExportedTypeMetadata<{baseClassOrInterfaceName}, {attributeName}>[]
-            {{
-                {string.Join(@",
-                ", exportedItemList)}
-            }}";
-
-            var methodCode = $@"ExportedTypeMetadata<{baseClassOrInterfaceName}, {attributeName}>[] ICompileTimeTypesExporter<{baseClassOrInterfaceName}, {attributeName}>.ExportTypes()
-        {{
-            return {arrayExpression};
-        }}";
-
-            exportedMethodCodes.Add(methodCode);
-
-            exportedInterfaces.Add($@"ICompileTimeTypesExporter<{baseClassOrInterfaceName}, {attributeName}>");
-        }
-
-        var code = $@"using dotnetCampus.Telescope;
-
-namespace dotnetCampus.Telescope
-{{
-    public partial class __AttributedTypesExport__ : {string.Join(", ", exportedInterfaces)}
-    {{
-        {string.Join(@"
-        ", exportedMethodCodes)}
-    }}
-}}";
-        code = FormatCode(code);
-        return code;
-    }
-
-    /// <summary>
-    /// 格式化代码。
-    /// </summary>
-    /// <param name="sourceCode">未格式化的源代码。</param>
-    /// <returns>格式化的源代码。</returns>
-    internal static string FormatCode(string sourceCode)
-    {
-        var rootSyntaxNode = CSharpSyntaxTree.ParseText(sourceCode).GetRoot();
-        return rootSyntaxNode.NormalizeWhitespace().SyntaxTree.GetText().ToString();
-    }
-
-    private static string TypeInfoToFullName(TypeInfo typeInfo)
-    {
-        ITypeSymbol typeSymbol = typeInfo.Type;
-        return TypeSymbolToFullName(typeSymbol);
-    }
-
-    private static string TypeSymbolToFullName(ITypeSymbol typeSymbol)
-    {
-        // 带上 global 格式的输出 FullName 内容
-        var symbolDisplayFormat = new SymbolDisplayFormat
-        (
-            // 带上命名空间和类型名
-            SymbolDisplayGlobalNamespaceStyle.Included,
-            // 命名空间之前加上 global 防止冲突
-            SymbolDisplayTypeQualificationStyle
-                .NameAndContainingTypesAndNamespaces
-        );
-
-        return typeSymbol?.ToDisplayString(symbolDisplayFormat);
-    }
-}
-
-readonly struct MarkClassParseResult
-{
-    public MarkClassParseResult(AssemblyCandidateClassParseResult classParseResult,
-        MarkExportAttributeParseResult markExportAttributeParseResult)
-    {
-        ClassParseResult = classParseResult;
-        MarkExportAttributeParseResult = markExportAttributeParseResult;
-    }
-
-    public AssemblyCandidateClassParseResult ClassParseResult { get; }
-
-    public MarkExportAttributeParseResult MarkExportAttributeParseResult { get; }
-}
-
-readonly struct AssemblyCandidateClassParseResult
-{
-    public AssemblyCandidateClassParseResult(INamedTypeSymbol typeInfo, ImmutableArray<AttributeData> attributes, GeneratorSyntaxContext generatorSyntaxContext)
-    {
-        TypeInfo = typeInfo;
-        Attributes = attributes;
-        GeneratorSyntaxContext = generatorSyntaxContext;
-        Success = true;
-    }
-
-    public AssemblyCandidateClassParseResult()
-    {
-        Success = false;
-        TypeInfo = default;
-        Attributes = default;
-        GeneratorSyntaxContext = default;
-    }
-
-    public INamedTypeSymbol TypeInfo { get; }
-
-    public ImmutableArray<AttributeData> Attributes { get; }
-
-    public GeneratorSyntaxContext GeneratorSyntaxContext { get; }
-
-    public bool Success { get; }
-}
-
-readonly struct MarkExportAttributeParseResult
-{
-    public MarkExportAttributeParseResult(bool success, TypeInfo baseClassOrInterfaceTypeInfo,
-        TypeInfo attributeTypeInfo)
-    {
-        Success = success;
-        BaseClassOrInterfaceTypeInfo = baseClassOrInterfaceTypeInfo;
-        AttributeTypeInfo = attributeTypeInfo;
-    }
-
-    public bool Success { get; }
-    public TypeInfo BaseClassOrInterfaceTypeInfo { get; }
-
-    public TypeInfo AttributeTypeInfo { get; }
-}
